+++ swift.js
@@ -51,17 +51,19 @@ module.exports = (config) => {
               password : _storage.password
             }
           }
-        },
-        scope : {
-          project : {
-            domain : {
-              id : 'default'
-            },
-            name : _storage.tenantName
-          }
         }
       }
     };
+    if (_storage.tenantName) {
+      _json.auth.scope = {
+        project : {
+          domain : {
+            id : 'default'
+          },
+          name : _storage.tenantName
+        }
+      }
+    }
 
     rock.concat({
       url    : `${_storage.authUrl}/auth/tokens`,
@@ -172,13 +174,13 @@ module.exports = (config) => {
    *
    * @param {string} container Container name
    * @param {string} filename file to store
-   * @param {string|Buffer} localPathOrBuffer absolute path to the file
+   * @param {string|Buffer|Function} localPathOrBuffer absolute path to the file
    * @param {Object} options [OPTIONAL]: { headers: {}, queries: {} } List of query parameters and headers: https://docs.openstack.org/api-ref/object-store/?expanded=create-or-replace-object-detail#create-or-replace-object
    * @param {function} callback function(err):void = The `err` is null by default, return an object if an error occurs.
    * @returns {void}
    */
   function uploadFile (container, filename, localPathOrBuffer, options, callback) {
-    let readStream = Buffer.isBuffer(localPathOrBuffer) === true ? localPathOrBuffer : () => { return fs.createReadStream(localPathOrBuffer) } ;
+    let readStream = Buffer.isBuffer(localPathOrBuffer) === true || typeof localPathOrBuffer === 'function' ? localPathOrBuffer  : () => { return fs.createReadStream(localPathOrBuffer) } ;
 
     const arrayArguments = [...arguments];